# Use StaticArrays for small matrices and vectors to avoid heap allocations.