--- a/REQUIRE
+++ b/REQUIRE
@@ -8,3 +8,4 @@ ForwardDiff 0.2
 JLD
 Optim 0.6
 WCS 0.1.3
+StaticArrays