--- a/src/DeterministicVI.jl
+++ b/src/DeterministicVI.jl
@@ -14,6 +14,7 @@ import ..Log
 using ..Transform
 import DataFrames
 import Optim
+using StaticArrays
 
 export ElboArgs