--- a/src/Model.jl
+++ b/src/Model.jl
@@ -1,5 +1,7 @@
 module Model
 
+using StaticArrays
+
 # parameter types
 export Image, TiledImage, ImageTile,
        SkyPatch, PsfComponent,