--- a/src/model/param_set.jl
+++ b/src/model/param_set.jl
@@ -121,21 +121,20 @@ getids(::Type{UnconstrainedParams}) = ids_free
 length(::Type{UnconstrainedParams}) =  6 + 2*Ia + 2*(B-1)*Ia + (D-1)*Ia + Ia-1
 
 # Parameters for a representation of the PSF
-type PsfParams <: ParamSet
-    mu::Vector{Int}
+immutable PsfParams <: ParamSet
+    mu::UnitRange{Int}
     e_axis::Int
     e_angle::Int
     e_scale::Int
     weight::Int
 
     function PsfParams()
-      new([1, 2], 3, 4, 5, 6)
+      new(1:2, 3, 4, 5, 6)
     end
 end
 const psf_ids = PsfParams()
 getids(::Type{PsfParams}) = psf_ids
 length(::Type{PsfParams}) = 6
-
 
 # define length(value) in addition to length(Type) for ParamSets
 length{T<:ParamSet}(::T) = length(T)