--- a/src/model/psf_model.jl
+++ b/src/model/psf_model.jl
@@ -16,15 +16,15 @@ Attributes:
 """
 immutable PsfComponent
     alphaBar::Float64  # TODO: use underscore
-    xiBar::Vector{Float64}
-    tauBar::Matrix{Float64}
+    xiBar::SVector{2,Float64}
+    tauBar::SMatrix{2,2,Float64,4}
 
-    tauBarInv::Matrix{Float64}
+    tauBarInv::SMatrix{2,2,Float64,4}
     tauBarLd::Float64
 
-    function PsfComponent(alphaBar::Float64, xiBar::Vector{Float64},
-                          tauBar::Matrix{Float64})
-        new(alphaBar, xiBar, tauBar, tauBar^-1, logdet(tauBar))
+    function PsfComponent(alphaBar::Float64, xiBar::SVector{2,Float64},
+                          tauBar::SMatrix{2,2,Float64,4})
+        new(alphaBar, xiBar, tauBar, inv(tauBar), log(det(tauBar)))
     end
 end