--- a/test/Synthetic.jl
+++ b/test/Synthetic.jl
@@ -17,35 +17,27 @@ function wrapped_poisson(rate::Float64)
 end
 
 
-function get_patch(the_mean::Vector{Float64}, H::Int, W::Int)
-    const radius = 50
+function get_patch(the_mean::SVector{2,Float64}, H::Int, W::Int)
+    radius = 50
     hm = round(Int, the_mean[1])
     wm = round(Int, the_mean[2])
     w11 = max(1, wm - radius):min(W, wm + radius)
     h11 = max(1, hm - radius):min(H, hm + radius)
-    return(w11, h11)
+    return (w11, h11)
 end
 
 
 function write_gaussian(the_mean, the_cov, intensity, pixels;
                         expectation=false)
-    the_precision = the_cov^-1
-    c = det(the_precision)^.5 / 2pi
-    y = Array(Float64, 2)
+    the_precision = inv(the_cov)
+    c = sqrt(det(the_precision)) / 2pi
 
     H, W = size(pixels)
     w_range, h_range = get_patch(the_mean, H, W)
 
-    function matvec222(mat::Matrix, vec::Vector)
-        # x' A x in a slightly more efficient form.
-        (mat[1,1] * vec[1] + mat[1,2] * vec[2]) * vec[1] +
-                (mat[2,1] * vec[1] + mat[2,2] * vec[2]) * vec[2]
-    end
-
     for w in w_range, h in h_range
-        y[1] = the_mean[1] - h
-        y[2] = the_mean[2] - w
-        ypy = matvec222(the_precision, y)
+        y = @SVector [the_mean[1] - h, the_mean[2] - w] # Maybe not hard code Float64
+        ypy = dot(y,  the_precision * y)
         pdf_hw = c * exp(-0.5 * ypy)
         pixel_rate = intensity * pdf_hw
         pixels[h, w] += expectation ? pixel_rate : wrapped_poisson(pixel_rate)