--- a/test/test_elbo.jl
+++ b/test/test_elbo.jl
@@ -3,6 +3,7 @@ using Celeste: DeterministicVI, SensitiveFloats
 using Base.Test
 using Distributions
 using DerivativeTestUtils
+using StaticArrays
 
 ######################################
 # Helper functions
@@ -333,7 +334,7 @@ function test_tiny_image_tiling()
   # point with a narrow psf.
 
   blob0 = SampleData.load_stamp_blob(datadir, "164.4311-39.0359_2kpsf");
-  pc = PsfComponent(1./3, zeros(2), 1e-4 * eye(2));
+  pc = PsfComponent(1/3, zeros(SVector{2,Float64}), 1e-4 * eye(SMatrix{2,2,Float64,4}));
   trivial_psf = [pc, pc, pc]
   pixels = ones(100, 1) * 12
   pixels[98:100, 1] = [1e3, 1e4, 1e5]