@jeff-regier 2016-10-14
Looks great, thanks for reducing the memory allocations.